:hammer: extract io functions from opt.go file to io.go
