+++ internal/opt.go
@@ -3,10 +3,7 @@ package internal
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
-	"strings"
-	"text/template"
 
 	"github.com/hellflame/argparse"
 )
@@ -15,11 +12,6 @@ type Options struct {
 	Template   *os.File
 	Output     *os.File
 	DataParser DataParser
-}
-
-type Context struct {
-	Template *template.Template
-	Data     map[string]interface{}
 }
 
 func NewOptions(args []string) (opts Options, err error) {
@@ -76,45 +68,4 @@ func NewOptions(args []string) (opts Options, err error) {
 	}
 
 	return
-}
-
-func (opts Options) ReadInputFiles() (ctx Context, err error) {
-	templateContent, err := ioutil.ReadAll(opts.Template)
-	if err != nil {
-		return
-	}
-
-	ctx.Template, err = template.New("template").Parse(string(templateContent))
-	if err != nil {
-		return
-	}
-
-	data, err := opts.DataParser.GetData()
-	if err != nil {
-		return
-	}
-
-	env := getEnvironment()
-
-	ctx.Data = map[string]interface{}{
-		"Data": data,
-		"Env":  env,
-	}
-
-	return ctx, err
-}
-
-func getEnvironment() map[string]string {
-	var data = map[string]string{}
-
-	for _, env := range os.Environ() {
-		var splitted = strings.SplitN(env, "=", 2)
-		data[splitted[0]] = splitted[1]
-	}
-
-	return data
-}
-
-func (opts Options) WriteOutput(context Context) (err error) {
-	return context.Template.Execute(opts.Output, context.Data)
 }
